+++ lib/gokart-core/src/types.rs
@@ -2,5 +2,5 @@ pub type Int = i64;
 pub type Double = f64;
 pub type Str = String;
 
-pub type Tag = usize;
-pub type Label = usize;
+pub type Tag = u64;
+pub type Label = u64;
